requirements.txt
streamlit>=1.35.0
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.2.0